fix: Address PR #53 review feedback
- NetworkPolicy: Use matchExpressions with Exists operator to allow ALL
  KServe InferenceServices, not just mistral-7b-development
- S3 Storage: Enable SSL verification (S3_VERIFY_SSL=1) for security
- Migration: Add explanatory comment for partial unique index in pull_jobs
- CLI: Refactor PullModelToCache to internal/cache package for reusability
- API Router: Replace startsWithOrganizationsMe with noV1PrefixPaths slice
  for better maintainability when adding future path exceptions
